# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Retro Board API & MCP Node",
    "version": "1.0.0",
    "description": "Unified API and MCP tool documentation for the Self-hosted Retrospective Board. This specification covers both traditional REST endpoints and the Agentic Workspace capabilities."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {},
    "schemas": {
      "MCP_Tool": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "inputSchema": {
            "type": "object"
          }
        }
      },
      "MCP_Resource": {
        "type": "object",
        "properties": {
          "uri": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/api/boards": {
      "get": {
        "summary": "List all boards",
        "description": "Retrieve a flat list of all boards in the workspace.",
        "responses": {
          "200": {
            "description": "A list of boards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "group_id": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/groups": {
      "get": {
        "summary": "List all board groups",
        "description": "Retrieve all workspace groups with their assigned boards.",
        "responses": {
          "200": {
            "description": "A list of board groups"
          }
        }
      }
    },
    "/api/boards/{boardId}": {
      "get": {
        "summary": "Get full board state",
        "description": "Retrieve all columns, cards, and metadata for a specific board.",
        "parameters": [
          {
            "in": "path",
            "name": "boardId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full board JSON"
          },
          "404": {
            "description": "Board not found"
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "summary": "Get app settings",
        "responses": {
          "200": {
            "description": "App settings JSON"
          }
        }
      },
      "post": {
        "summary": "Update app settings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings JSON"
          }
        }
      }
    },
    "/api/upload": {
      "post": {
        "summary": "Upload an image",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image URL JSON"
          }
        }
      }
    },
    "/mcp": {
      "get": {
        "summary": "MCP SSE Handshake",
        "description": "Establishes a Server-Sent Events (SSE) connection for the Model Context Protocol.",
        "responses": {
          "200": {
            "description": "SSE stream established"
          }
        }
      },
      "post": {
        "summary": "MCP JSON-RPC Gateway",
        "description": "Unified entry point for all MCP tool calls and resource discoveries. Supports both stateful (via sessionId) and stateless (direct) requests.",
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "schema": {
              "type": "string"
            },
            "description": "Required for stateful SSE sessions."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                  },
                  "method": {
                    "type": "string"
                  },
                  "params": {
                    "type": "object"
                  },
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response"
          }
        }
      }
    }
  },
  "tags": []
}